Filter the most common problematic publisher URLs
Whether the links go to the record/abstract or the PDF, those tend
to be unreliable (hybrid articles often become paywalled agani) and
have been often removed on the English Wikipedia (including for being
perceived as duplicates of the DOI link).

The open archives also tend to be more user-friendly, with less
privacy-invasive JavaScript and so on.